\name{fit.learning.curve}
\alias{fit.learning.curve}
\title{Performs a power-law fit to the learning curve}
\usage{
  fit.learning.curve(stats)
}
\arguments{
  \item{stats}{a learning curve summary statistics object,
  as returned by \code{lc.stats}}
}
\value{
  a lm object for the power-law fit
}
\description{
  Performs a power-law fit to the learning curve
}

