\name{get.correct.from.opps}
\alias{get.correct.from.opps}
\title{The number of correct responses at each opportunity}
\usage{
  get.correct.from.opps(opps)
}
\arguments{
  \item{opps}{an opportunity table}
}
\value{
  a numeric vector with length equal to the number of rows
  in opps, giving the number of correct responses at each
  opportunity
}
\description{
  Calculates the number of correct responses at each
  opportunity across all the students in an opportunity
  table
}

