\name{get.total.from.opps}
\alias{get.total.from.opps}
\title{The total number of responses at each opportunity}
\usage{
  get.total.from.opps(opps)
}
\arguments{
  \item{opps}{an opportunity table}
}
\value{
  a numeric vector with length equal to the number of rows
  in opps, giving the total number of responses at each
  opportunity
}
\description{
  Calculates the total number of responses at each
  opportunity across all the students in an opportunity
  table
}

