\name{lc.title}
\alias{lc.title}
\title{Format a learning curve plot title}
\usage{
  lc.title(sk)
}
\arguments{
  \item{sk}{the skill for which to generate a title}
}
\value{
  a string with a nicely formatted title
}
\description{
  Utility function to produce a title suitable for a
  learning curve of the given skill.
}

